requirements.txt
Flask
Flask-Login
dash
dash-bootstrap-components
authlib
gunicorn
gspread
google-auth
google-auth-oauthlib
google-auth-httplib2
pandas